+++ src/Generics/Deriving/Copoint.hs
@@ -5,11 +5,16 @@
 
 #if __GLASGOW_HASKELL__ >= 701
 {-# LANGUAGE DefaultSignatures #-}
-{-# LANGUAGE Safe #-}
 #endif
 
 #if __GLASGOW_HASKELL__ >= 705
 {-# LANGUAGE PolyKinds #-}
+#endif
+
+#if __GLASGOW_HASKELL__ >= 710
+{-# LANGUAGE Safe #-}
+#elif __GLASGOW_HASKELL__ >= 701
+{-# LANGUAGE Trustworthy #-}
 #endif
 
 module Generics.Deriving.Copoint (
@@ -30,6 +35,12 @@ import           Data.Monoid (Dual)
 import qualified Data.Monoid as Monoid (Sum)
 
 import           Generics.Deriving.Base
+
+#if MIN_VERSION_base(4,6,0)
+import           Data.Ord (Down)
+#else
+import           GHC.Exts (Down)
+#endif
 
 #if MIN_VERSION_base(4,8,0)
 import           Data.Functor.Identity (Identity)
@@ -125,6 +136,9 @@ instance GCopoint f => GCopoint (Alt f) where
 instance GCopoint (Arg a) where
   gcopoint = gcopointdefault
 #endif
+
+instance GCopoint Down where
+  gcopoint = gcopointdefault
 
 instance GCopoint Dual where
   gcopoint = gcopointdefault