+++ src/Generics/Deriving/Functor.hs
@@ -7,7 +7,6 @@
 
 #if __GLASGOW_HASKELL__ >= 701
 {-# LANGUAGE DefaultSignatures #-}
-{-# LANGUAGE Safe #-}
 #endif
 
 #if __GLASGOW_HASKELL__ >= 705
@@ -16,6 +15,12 @@
 
 #if __GLASGOW_HASKELL__ >= 708
 {-# LANGUAGE EmptyCase #-}
+#endif
+
+#if __GLASGOW_HASKELL__ >= 710
+{-# LANGUAGE Safe #-}
+#elif __GLASGOW_HASKELL__ >= 701
+{-# LANGUAGE Trustworthy #-}
 #endif
 
 module Generics.Deriving.Functor (
@@ -39,6 +44,12 @@ import           Generics.Deriving.Base
 
 #if MIN_VERSION_base(4,4,0)
 import           Data.Complex (Complex)
+#endif
+
+#if MIN_VERSION_base(4,6,0)
+import           Data.Ord (Down)
+#else
+import           GHC.Exts (Down)
 #endif
 
 #if MIN_VERSION_base(4,7,0)
@@ -154,6 +165,9 @@ instance GFunctor Complex where
 #endif
 
 instance GFunctor (Const m) where
+  gmap = gmapdefault
+
+instance GFunctor Down where
   gmap = gmapdefault
 
 instance GFunctor Dual where